refactor(downloads): extract DownloadTask into its own domain model file
